Add further effects processing
More extensive changes have been made to the Max Patch to ensure that the processing is sophisticated. Minor changes have been made to the server to allow for secondary emotions to hold less weight in their mix prominence. This will likely be done through quieter volume, though it is also possible that this will also hapeen via reducing the mix amount of effects