Improve emotion wheel interactivity
These changes will help to make the emotion wheel a bit more interactive and engaging. There is now an add and a next button (the former for adding an emotion to the soundscape, the latter for continuing to the next page). To make the process more interesting, it would be good to make the process happen in turns, but this is something that can be implemented later.