Globally scale Max Patch volume
The loudness of the output is contingent on the number of sounds present. Not only do sounds smoothly fade in and out, but the volume multiplier associated with sounds playing/not playing gradually transitions rather than being rough. Subtle changes like this will help the sound effects to be more subtle and seamless